feat: add build hooks for custom config
