+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools>=61", "wheel", "build"]
+requires = ["setuptools>=64", "wheel", "build"]
 build-backend = "setuptools.build_meta"
 
 [project]
@@ -9,7 +9,7 @@ description = "QC Lab: a python package for quantum-classical modeling."
 readme = "README.rst"
 requires-python = ">=3.6"
 license = { file = "LICENSE" }
-authors = [ { name="Tempelaar Team", email="roel.tempelaar@northwestern.edu" } ]
+authors = [{ name = "Tempelaar Team", email = "roel.tempelaar@northwestern.edu" }]
 classifiers = [
     "Development Status :: 3 - Alpha",
     "Intended Audience :: Education",
@@ -19,12 +19,19 @@ classifiers = [
     "Operating System :: Linux :: Linux",
     "License :: OSI Approved :: Apache Software License",
 ]
+dependencies = [
+    "numpy",
+    "tqdm",
+    "h5py",
+    "matplotlib",
+    "numba",
+]
 
 [project.optional-dependencies]
 tests = [
     "pytest",
     "mpi4py",
-    ]
+]
 docs = [
     "sphinx",
     "pydata-sphinx-theme",
@@ -33,3 +40,10 @@ docs = [
     "sphinxcontrib-mermaid",
     "graphviz",
 ]
+
+[tool.setuptools]
+packages = ["find:"]
+
+[tool.setuptools.cmdclass]
+develop = "qc_lab._build_hooks:develop"
+build_py = "qc_lab._build_hooks:build_py"
